requirements.txt
certifi
chardet
elasticsearch
python-logstash
idna
numpy
pandas
psycopg2
pymc
python-dateutil
pytz
requests
scikit-learn
scipy
six
sklearn
SQLAlchemy
urllib3
